・関数名の変更、コードの位置変更、処理の関数化
